add Employees controller
