+++ WebStore/Views/Employees/EmployeeDetail.cshtml
