+++ WebStore/Views/Home/Index.cshtml
@@ -12,6 +12,6 @@
 </head>
 <body>
     <h3>Практическое занятие. Урок 1.</h3>
-    <a href="~/Home/Employees">Сотрудники</a>
+    <a asp-controller="Employees">Сотрудники</a>
 </body>
 </html>
